+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 ## 1.2.0-beta1
 * Bastion Hosts: Create bastion hosts for accessing resources on a virtual network.
 * Gateway: Add VPN Client configuration
+* Service Bus: Set message TTL 
 
 ## 1.1.0
 * AKS: Basic AKS support